Refine how profile images are set
Instead of using React's `this.state` properties, we add a property on
the invocation of `<User />` to pass the image on, because `User` is a
child of `PlaylistList`.

Takeaway: To pass data to children, use `this.props`

We add the ability for the header component to listen for a new profile
picture on `'tm:user'` events (by merge the whole event `e` into it's
state).

Takeaway: To communicate with non-ancestors, use events attached to
global elements (i.e., `body`).

Also, a small nit: I hardcode the placeholder URL in the frontend code,
removing it from the backend code for consistency.